% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAME.R
\name{calcAME}
\alias{calcAME}
\title{Calculate AME}
\usage{
calcAME(x, groupvar, timevar, what, vfr, dat)
}
\arguments{
\item{x}{Character string. Treatment variable.}

\item{groupvar}{Grouping variable}

\item{timevar}{Time variable}

\item{what}{"Mu" or "Sigma"}

\item{vfr}{gamlss output}

\item{dat}{Dataframe}
}
\value{
List of length 2. Element 1 returns AME by group and time. Elements 2 returns the AME by time.
}
\description{
[...]
}
\examples{
data(incdat)
vfr <- gamlss()
AME_mu <- calcAME("x", groupvar=SES, timevar=year, what="mu", vfr, incdat)

}
\author{
Benjamin Rosche <benjamin.rosche@gmail.com>
}
