% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_formulas.R
\name{compute_delta_B}
\alias{compute_delta_B}
\title{Compute treatment effect on between-group inequality}
\usage{
compute_delta_B(pi, mu0, beta, ystat)
}
\arguments{
\item{pi}{Group shares among treated}

\item{mu0}{Baseline group means mu_j(0)}

\item{beta}{Treatment effects on mean}

\item{ystat}{"Var" or "CV2"}
}
\value{
Scalar delta_B^D
}
\description{
Compute treatment effect on between-group inequality
}
\keyword{internal}
