% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cps_sample}
\alias{cps_sample}
\title{CPS sample data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 65562 rows and 32 columns.
}
\source{
\url{https://cps.ipums.org/cps/}
}
\usage{
data(cps_sample)
}
\description{
A sample from the Current Population Survey containing women's earnings
data with motherhood status, used to demonstrate the causal variance
decomposition.
}
\examples{
data(cps_sample)
}
\references{
Sarah Flood, Miriam King, Renae Rodgers, Steven Ruggles,
J. Robert Warren and Michael Westberry. Integrated Public Use Microdata
Series, Current Population Survey: Version 9.0 [dataset].
Minneapolis, MN: IPUMS, 2021.
\doi{10.18128/D030.V9.0}
}
\keyword{datasets}
