% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_formulas.R
\name{CV2B_pi}
\alias{CV2B_pi}
\title{Between-group CV^2 (pi-based)}
\usage{
CV2B_pi(pi, mu)
}
\arguments{
\item{pi}{Numeric vector of group shares}

\item{mu}{Numeric vector of group means}
}
\value{
Scalar between-group CV^2: sum(pi * (mu - grand_mean)^2) / grand_mean^2
}
\description{
Between-group CV^2 (pi-based)
}
\keyword{internal}
