% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{.build_longit_subcomponents}
\alias{.build_longit_subcomponents}
\title{Build and print cross-sectional sub-component table from longitudinal results}
\usage{
.build_longit_subcomponents(
  results,
  ref,
  cross_se = NULL,
  lambda_tests = NULL,
  beta_tests = NULL
)
}
\description{
Build and print cross-sectional sub-component table from longitudinal results
}
\keyword{internal}
