% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{.build_outcome_dist_data}
\alias{.build_outcome_dist_data}
\title{Build outcome dist data (Control + Treated per group)}
\usage{
.build_outcome_dist_data(groups, mu0, sigma0, beta, lambda, trim = 0.995)
}
\description{
Build outcome dist data (Control + Treated per group)
}
\keyword{internal}
