% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{.build_te_dist_data}
\alias{.build_te_dist_data}
\title{Build treatment effect dist data (one density per group)}
\usage{
.build_te_dist_data(groups, mu0, sigma0, beta, lambda, trim = 0.995)
}
\description{
Build treatment effect dist data (one density per group)
}
\keyword{internal}
