% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{.compute_wibe_from_d}
\alias{.compute_wibe_from_d}
\title{Recompute wibe summary from raw data (for bootstrap)}
\usage{
.compute_wibe_from_d(d)
}
\description{
Recompute wibe summary from raw data (for bootstrap)
}
\keyword{internal}
