% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{.get_all_times}
\alias{.get_all_times}
\title{Get sorted list of all times (ref + result times)}
\usage{
.get_all_times(results, ref)
}
\description{
Get sorted list of all times (ref + result times)
}
\keyword{internal}
