% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_formulas.R
\name{.ineq_se_Gini}
\alias{.ineq_se_Gini}
\title{Linearised SE for weighted Gini (Monti 1991 / Deville 1999)}
\usage{
.ineq_se_Gini(y, w = NULL, val)
}
\description{
Linearised SE for weighted Gini (Monti 1991 / Deville 1999)
}
\keyword{internal}
