% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_formulas.R
\docType{data}
\name{.ineq_stat_registry}
\alias{.ineq_stat_registry}
\title{Registry mapping stat names to functions}
\format{
An object of class \code{list} of length 5.
}
\usage{
.ineq_stat_registry
}
\description{
Registry mapping stat names to functions
}
\keyword{internal}
