% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{.print_causal_block}
\alias{.print_causal_block}
\title{Print a single causal per-time block}
\usage{
.print_causal_block(r, se = NULL)
}
\description{
Print a single causal per-time block
}
\keyword{internal}
