% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{.print_desc_block}
\alias{.print_desc_block}
\title{Print a single descriptive per-time block}
\usage{
.print_desc_block(row, ystat = "Var")
}
\description{
Print a single descriptive per-time block
}
\keyword{internal}
