% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{.print_indented_df}
\alias{.print_indented_df}
\title{Print a data.frame with indentation}
\usage{
.print_indented_df(df, indent = 4, digits = 4)
}
\description{
Print a data.frame with indentation
}
\keyword{internal}
