% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{.print_line}
\alias{.print_line}
\title{Print a single line: label + value + optional SE}
\usage{
.print_line(label, value, se = NULL, indent = 4, pad = 52)
}
\description{
Print a single line: label + value + optional SE
}
\keyword{internal}
