% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{.print_val_with_se}
\alias{.print_val_with_se}
\title{Print estimate with SE (legacy, used by cross-sectional)}
\usage{
.print_val_with_se(label, estimate, se)
}
\description{
Print estimate with SE (legacy, used by cross-sectional)
}
\keyword{internal}
