% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{incdat}
\alias{incdat}
\title{Income data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3000 rows and 7 columns.
}
\usage{
data(incdat)
}
\description{
A simulated income dataset for demonstrating the variance decomposition
functions.
}
\examples{
data(incdat)
}
\keyword{datasets}
