% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.ineqx_shapley}
\alias{plot.ineqx_shapley}
\title{Plot Shapley values}
\usage{
\method{plot}{ineqx_shapley}(x, type = "four", ...)
}
\arguments{
\item{x}{An \code{ineqx_shapley} object}

\item{type}{Character: \code{"four"} or \code{"six"}}

\item{...}{Additional arguments (currently unused)}
}
\value{
A ggplot2 object
}
\description{
Plot Shapley values
}
