% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShinyExample.R
\name{runShinyExample}
\alias{runShinyExample}
\title{runShinyExample}
\usage{
runShinyExample()
}
\description{
Run to view shiny app demonstrating the ineqx package.
}
\examples{
runShinyExample()
}
\author{
Benjamin Rosche <benjamin.rosche@gmail.com>
}
