% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ineqx.R
\docType{data}
\name{theme_ineqx}
\alias{theme_ineqx}
\title{ineqx ggplot2 theme}
\format{
An object of class \code{theme} (inherits from \code{ggplot2::theme}, \code{gg}, \code{S7_object}) of length 144.
}
\usage{
theme_ineqx
}
\description{
A custom ggplot2 theme for consistent styling of ineqx plots.
Use as \code{+ theme_ineqx} in ggplot2 pipelines.
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(wt, mpg)) + geom_point() + theme_ineqx

}
\keyword{datasets}
