% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wibe.R
\name{wibe}
\alias{wibe}
\title{Descriptive within/between decomposition}
\usage{
wibe(
  dat,
  y = NULL,
  groupvar = NULL,
  timevar = NULL,
  smoothDat = F,
  rel = F,
  long = F
)
}
\arguments{
\item{dat}{Dataframe}

\item{y}{Dependent variable}

\item{groupvar}{To create within/between division}

\item{timevar}{To analyze change over time}

\item{smoothDat}{TRUE|FALSE. Smooth data?}

\item{rel}{FALSE or numeric time}

\item{long}{TRUE|FALSE. Long instead of wide output}
}
\value{
Returns a list with two elements - "by group and time" and "by time"
}
\description{
[...]
}
\examples{
data(incdat)
wibe1 <- wibe(dat, y=inc, groupvar=SES, timevar=year)

}
\author{
Benjamin Rosche <benjamin.rosche@gmail.com>
}
