% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cps-data.R
\docType{data}
\name{cps_sample}
\alias{cps_sample}
\title{CPS sample}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 102176 rows and 31 columns.
}
\source{
\href{https://cps.ipums.org/cps/}{CPS Website}
}
\usage{
data(cps_sample)
}
\description{
CPS sample
}
\examples{
data(cps_sample)
}
\references{
Sarah Flood, Miriam King, Renae Rodgers, Steven Ruggles, J. Robert Warren and Michael Westberry.
Integrated Public Use Microdata Series, Current Population Survey: Version 9.0 [dataset]. Minneapolis, MN: IPUMS, 2021.
(\href{https://doi.org/10.18128/D030.V9.0}{https://doi.org/10.18128/D030.V9.0})
}
\keyword{datasets}
