% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{incdat}
\alias{incdat}
\title{Income data}
\format{
A data frame with x rows and y variables:
\describe{
  \item{Col1}{description of Col1}
...
}
}
\usage{
data(incdat)
}
\description{
...
}
\examples{
data(incdat)
}
\keyword{datasets}
