% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ineqx.R
\name{ineqx}
\alias{ineqx}
\title{Variance function regression + decomposition of the variance and sq. coefficient of variation}
\usage{
ineqx(
  x,
  y,
  ystat = "CV2",
  controls = NULL,
  groupvar = NULL,
  timevar = NULL,
  timesp = T,
  cf = 1,
  dat
)
}
\arguments{
\item{x}{Explanatory variable}

\item{y}{Dependent variable}

\item{ystat}{"Var" or "CV2"}

\item{controls}{Character vector with additional control variables}

\item{groupvar}{To create within/between division}

\item{timevar}{To analyze change over time}

\item{cf}{Either Flase or numeric time}

\item{dat}{Dataframe}
}
\value{
Returns a list: dW, dB, dD, dT
}
\description{
The ineqx package implements Rosche (2021). [...]
}
\examples{
data(incdat)
decomp1 <- ineqx(child, inc, ystat="CV2", groupvar=SES, timevar=year, cf=1, incdat)

}
\author{
Benjamin Rosche <benjamin.rosche@gmail.com>
}
