% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ineqx.R
\name{ineqx}
\alias{ineqx}
\title{Descriptive and causal variance decompositions}
\usage{
ineqx(
  treat = NULL,
  post = NULL,
  y,
  ystat = "Var",
  group = NULL,
  time = NULL,
  weights = NULL,
  controls = NULL,
  decomp = "post",
  ref = NULL,
  AME_mu = NULL,
  AME_sigma = NULL,
  dat
)
}
\arguments{
\item{treat}{Character string. Treatment variable. Values must be 0/1}

\item{post}{Character string. Before/after variable. Values must be 0/1}

\item{y}{Character string. Dependent variable. Variable must be continuous.}

\item{ystat}{Character string. Either "Var" (default) or "CV2". Choose to analyze (the effect of x on) the variance or squared coefficient of variation.}

\item{group}{Character string. Variable must be categorical. Grouping variable to decompose variance into within- and between-group components.}

\item{time}{Character string. c.x with specify penalized splines, i.x will specify dummies. Time variable to analyze change over time.}

\item{weights}{Character string. Weight variable.}

\item{controls}{Character vector with additional control variables. E.g. c("c.age", "i.sex", ...)}

\item{decomp}{Character string. Either "post" (default) or "effect".}

\item{ref}{Number, vector, or list. Counterfactual reference point. See details.}

\item{AME_mu}{Dataframe with average marginal effects (Mu)}

\item{AME_sigma}{Dataframe with average marginal effects (Sigma)}

\item{dat}{Dataframe}
}
\value{
List with eight elements: "vars", "dMu", "dSigma", "dW", "dB", "dC", "dP", "dT". See details.
}
\description{
The ineqx package implements Rosche (202X). [...]
}
\details{
The main function of \code{ineqx}. [...]


\bold{Counterfactual reference point (ref)}

\code{ref} can either be a number, c(number, "variable name"), or list(n=c(...), mu=c(...), sigma=c(...), beta=c(...), lambda=c(...)). [...]

\bold{Return}

...
}
\examples{
data(incdat)
# Descriptive variance decomposition
decomp_descr <- ineqx(y="inc", group="group", time="i.year", ref=1, dat=incdat)
plot(decomp_descr, type="dT")

# Causal variance decomposition
decomp_treat <- ineqx(treat="x", post="t", y="inc", group="group", time="i.year", ref=1, dat=incdat)
plot(decomp_treat, type="dT")

}
\author{
Benjamin Rosche <benjamin.rosche@gmail.com>
}
