% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ineqx.R
\name{ineqx}
\alias{ineqx}
\title{Variance function regression + decomposition of the variance and sq. coefficient of variation}
\usage{
ineqx(
  x,
  y,
  xpv = c(0, 1),
  ystat = "CV2",
  controls = NULL,
  groupvar = NULL,
  timevar = NULL,
  cf = NULL,
  dat
)
}
\arguments{
\item{x}{Character string. Explanatory variable. c.x or i.x. c2.x will specify quadratic polynomials for x.}

\item{y}{Character string. Dependent variable. Can only be c.x}

\item{xpv}{Numeric vector of length 2. Choose the values of x that are used for prediction, e.g. c(0,1)}

\item{ystat}{Character string. Either "Var" or "CV2". Choose to analyze the effect of x on the variance or squared coefficient of variation}

\item{controls}{Character vector with additional control variables. E.g. c("c.age", "i.sex", ...)}

\item{groupvar}{Character string. Must be i.x. Grouping variable to decompose variance into within- and between-group components.}

\item{timevar}{Character string. c.x with specify penalized splines, i.x will specify dummies. Time variable to analyze change over time.}

\item{cf}{Number. Reference time.}

\item{dat}{Dataframe}
}
\value{
List with six elements: pred_mu, pred_sigma, dW, dB, dD, dT. See details.
}
\description{
The ineqx package implements Rosche (2021). [...]
}
\details{
The main function of \code{ineqx}. [...]
}
\examples{
data(incdat)
decomp1 <- ineqx(child, inc, ystat="CV2", groupvar=SES, timevar=year, cf=1, incdat)

}
\author{
Benjamin Rosche <benjamin.rosche@gmail.com>
}
