% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ineqx.R
\name{plot.ineqx}
\alias{plot.ineqx}
\title{plot function}
\usage{
\method{plot}{ineqx}(
  ineqx.out,
  type,
  yint = 1,
  xlab = NULL,
  ylab = NULL,
  llab = NULL,
  titl = NULL,
  xlim = NULL,
  hline = F,
  se = F,
  lowess = F,
  legend = T,
  sav = F
)
}
\arguments{
\item{ineqx.out}{ineqx.out object from ineqx()}

\item{type}{dMuSigma2|dB|dW|dD|dO|dT}

\item{yint}{1 or 2. 1: ..., 2: ...}

\item{sav}{}
}
\value{
Returns a ggplot2 object
}
\description{
[...]
}
\examples{
data(incdat)
i1 <- ineqx(incdat, by, inc, groupvar = SES, timevar = year)
plot(i1, type="dMuSigma2")

}
\author{
Benjamin Rosche <benjamin.rosche@gmail.com>
}
