% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ineqx.R
\name{plot.ineqx}
\alias{plot.ineqx}
\title{plot function}
\usage{
\method{plot}{ineqx}(
  ineqx.out,
  type,
  yint = 1,
  xlab = NULL,
  ylab = NULL,
  llab = NULL,
  titl = NULL,
  xlim = NULL,
  hline = F,
  se = F,
  lowess = F,
  legend = T,
  sav = F
)
}
\arguments{
\item{ineqx.out}{ineqx.out object from ineqx()}

\item{type}{Character string. Plot type. Choose from: dMuP, dMuT, dSigmaP, dSigmaT, dWP, dWT, dBP, dBT, dDP, dDT, dO, dT, dPA}

\item{yint}{Either 1 or 2. Choose reported effect value. 1: Effect in units of ystat. 2: Effect as change in percent since 'cf' (reference time)}

\item{xlab}{Character string. x-axis label.}

\item{ylab}{Vector of length 2 with character strings. y-axis label. c("y1lab", "y2lab")}

\item{llab}{Vector of length equal to the number of levels of groupvar. Group labels. c("grouplbl1", "grouplbl2", "grouplbl3", ...)}

\item{titl}{Character string of length 2. Plot titles. c("titl1", "titl2")}

\item{xlim}{Sequence of numbers.breaks at x-axis. seq(2000,2020, 5)}

\item{hline}{Number or FALSE. Horizontal line at specified value.}

\item{se}{Logical. Display standard errors?}

\item{lowess}{Logical. Display lowess smoother?}

\item{legend}{Logical. Display legend?}

\item{sav}{Logical. Save as .png in working directory?}
}
\value{
Returns a ggplot2 object
}
\description{
[...]
}
\examples{
data(incdat)
i1 <- ineqx(incdat, by, inc, groupvar = SES, timevar = year)
plot(i1, type="dMuSigma2")

}
\author{
Benjamin Rosche <benjamin.rosche@gmail.com>
}
