% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.ineqx_causal_cross}
\alias{plot.ineqx_causal_cross}
\title{Plot cross-sectional causal decomposition}
\usage{
\method{plot}{ineqx_causal_cross}(x, type = "bar", ...)
}
\arguments{
\item{x}{An \code{ineqx_causal_cross} object}

\item{type}{Character: \code{"bar"} for bar chart, \code{"group"} for
group-level contributions.}

\item{...}{Additional arguments (currently unused)}
}
\value{
A ggplot2 object
}
\description{
Plot cross-sectional causal decomposition
}
