% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.ineqx_causal_longit}
\alias{plot.ineqx_causal_longit}
\title{Plot longitudinal causal decomposition}
\usage{
\method{plot}{ineqx_causal_longit}(x, type = "four", ...)
}
\arguments{
\item{x}{An \code{ineqx_causal_longit} object}

\item{type}{Character:
\code{"four"} for four-component decomposition,
\code{"six"} for six-component decomposition,
\code{"wb"} for within/between totals.}

\item{...}{Additional arguments (currently unused)}
}
\value{
A ggplot2 object
}
\description{
Plot longitudinal causal decomposition
}
