% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.ineqx_desc}
\alias{plot.ineqx_desc}
\title{Plot descriptive variance decomposition}
\usage{
\method{plot}{ineqx_desc}(x, type = "deltas", ...)
}
\arguments{
\item{x}{An \code{ineqx_desc} object}

\item{type}{Character, plot type:
\code{"wibe"} for within/between levels over time,
\code{"deltas"} for change components over time (requires ref).}

\item{...}{Additional arguments (currently unused)}
}
\value{
A ggplot2 object
}
\description{
Plot descriptive variance decomposition
}
