% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ineqx.R
\docType{data}
\name{theme_ineqx}
\alias{theme_ineqx}
\title{theme_ineqx}
\format{
An object of class \code{theme} (inherits from \code{gg}) of length 93.
}
\usage{
theme_ineqx
}
\description{
ggplot theme used for the plotting
}
\examples{
theme_ineqx()
}
\author{
Benjamin Rosche <benjamin.rosche@gmail.com>
}
\keyword{datasets}
