% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wibe.R
\name{wibe}
\alias{wibe}
\title{Descriptive within/between decomposition}
\usage{
wibe(
  y = NULL,
  groupvar = NULL,
  timevar = NULL,
  dat,
  smoothDat = F,
  rel = F,
  long = F
)
}
\arguments{
\item{y}{Dependent variable}

\item{groupvar}{Grouping variable to decompose variance into within- and between-group components}

\item{timevar}{Time variable to analyze change over time}

\item{dat}{Dataframe}

\item{smoothDat}{Logical. Should data be smoothed?}

\item{rel}{Number or FALSE. Should values be reported relative to specified time?}

\item{long}{Logical. Should output be in long format?}
}
\value{
List of length 2. Element 1 returns the decomposition by group and time. Elements 2 returns the decomposition by time.
}
\description{
[...]
}
\examples{
data(incdat)
wibe1 <- wibe(dat, y=inc, groupvar=SES, timevar=year)

}
\author{
Benjamin Rosche <benjamin.rosche@gmail.com>
}
