% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wibe.R
\name{wibe}
\alias{wibe}
\title{Descriptive within/between decomposition}
\usage{
wibe(
  y = NULL,
  group = NULL,
  time = NULL,
  weights = NULL,
  ref = F,
  long = F,
  dat
)
}
\arguments{
\item{y}{Dependent variable}

\item{group}{Grouping variable to decompose variance into within- and between-group components}

\item{time}{Time variable to analyze change over time}

\item{weights}{Probability weights}

\item{ref}{Number or FALSE. Should values be reported in reference to a specific time?}

\item{long}{Logical. Should output be in long format?}

\item{dat}{Dataframe}

\item{smoothDat}{Logical. Should data be smoothed?}
}
\value{
List of length 2. Element 1 returns the decomposition by group and time. Elements 2 returns the decomposition by time.
}
\description{
[...]
}
\details{
...
}
\examples{
data(incdat)
wibe1 <- wibe(y="inc", group="SES", time="year", dat=dat1)

}
\author{
Benjamin Rosche <benjamin.rosche@gmail.com>
}
